[FIR] Move `RedundantCallOfConversionMethod` to the default checkers list
KT-77727

+++ compiler/testData/codegen/box/diagnostics/functions/tailRecursion/sum.kt
@@ -5,6 +5,6 @@ tailrec fun sum(x: Long, sum: Long): Long {
 
 fun box() : String {
     val sum = sum(1000000, 0)
-    if (sum != 500000500000.toLong()) return "Fail $sum"
+    if (sum != 500000500000L) return "Fail $sum"
     return "OK"
 }
